Enable Travis
Enable Travis CI builds using EnableTravisForRugArchiveTS.  This build
is a little different than other Rugs in that we want to build the
project itself to make sure it is working.  In fact, this is probably
the case for all generators.  So the standard Rug Travis CI build was
modified a bit to also run the Maven build.  The .travis.yml in the
generated project was also modified to be appropriate for the Maven
build.  The File extension was used rather than Travis or Yml because
neither of those currently provide the needed functionality and I am
too far down a rabbit hole as it is.