Add GitHub release
Configure Travis CI to do a GitHub release.  Update generator logic to
remove that from .travis.yml.  The Travis CLI adds double quotes to
lots if fields in the .travis.yml, so the replacement logic needed to
be smarter.

Generated language extension projects should npm publish on release,
so add that.